{{cookiecutter.package_name}}/requirements.txt
fabric3
flask >= 1.0
python-dotenv

